1.优化api。

+++ include/device/pin.h
@@ -75,8 +75,8 @@ struct mr_pin
 struct mr_pin_ops
 {
     int (*configure)(struct mr_pin *pin, int number, int mode);
-    int (*read)(struct mr_pin *pin, int number);
-    void (*write)(struct mr_pin *pin, int number, int value);
+    uint8_t (*read)(struct mr_pin *pin, int number);
+    void (*write)(struct mr_pin *pin, int number, uint8_t value);
 };
 
 /**
